Add filter
Well after a lot of messing around and a bunch of squashed
commits, we finally have a filter.

This is adapted from Crisp example by Robbert-VDH.
Someday I hope to design my own, but let's get this thing
working first.

The biquad filter doesn't rely on FFT. I also added a spectrum
calculator thing so that we can render a nice frequency plot.

Sadly the graphs are rendering in the wrong place.

Signed-off-by: Tobin Fitzthum <tobinf@protonmail.com>