Support multiple wave tables.
For presets to make sense, the wave table has to stay the same.
In order to support extending the wave table later, this patch
introduces backend support for multiple wave tables (known as
wave banks). Now we can add new banks while keeping the existing
ones the same.

Switching between wave banks is not yet suppoerted in the editor.

Signed-off-by: Tobin Fitzthum <tobinf@protonmail.com>