Switch from wave_warp to start and end indices
Hopefully this is more intuitive.

Also this will make it more likely for people to end up
with weird warping sounds. Previously warping would only
happen when wave_warp was not set to zero. Now it will
happen whenever the start and end are not equal, which
is basically always. If this ends up being annoying
we can switch back.

Signed-off-by: Tobin Fitzthum <tobinf@protonmail.com>