Switch preset menu to dropdown
This was supposed to fix a number of problems, but
it didn't. At least this is slightly nicer
than the scrolling thing (although we'll need to
add scrolling back into the mix if we have a lot of presets).

This re-introduced the issue with the textbox.

Behavior with the default value is still bad.
If we set "Default" as the default selection
everything will explode when you click load.

If we set some existing value, then the params
won't match the default.

Need to think about loading some default state automatically at the
beginning perhaps.

The preset menu also won't add any new presets to the dropdown
until the plugin is reloaded. THhis could possibly be fixed with
some kind of binding.

Signed-off-by: Tobin Fitzthum <tobinf@protonmail.com>